Move parsing logic to src/parsers and create a platform to add parsers - aspect-oriented programming.
